Dont require consumers to implement their own CommandOutputProvider
To get extension methods
