+++ source/Tests/Commands/HelpCommandFixture.cs
@@ -29,7 +29,7 @@ namespace Tests.Commands
 
             commandLocator = Substitute.For<ICommandLocator>();
             logger = new LoggerConfiguration().WriteTo.TextWriter(output).CreateLogger();
-            commandOutputProvider = new TestCommandOutputProvider(logger);
+            commandOutputProvider = new CommandOutputProvider("TestApp", new DefaultCommandOutputJsonSerializer(), logger);
             helpCommand = new HelpCommand(new Lazy<ICommandLocator>(() => commandLocator), commandOutputProvider);
         }
